evaluation code moved to separated module
